+++ programs/restaking/src/fund/events/token_deposited.rs
@@ -12,7 +12,7 @@ pub struct FundTokenDeposited {
     pub deposited_token_user_account: Pubkey,
 
     pub token_deposit_amount: u64,
-    pub token_amount_in_fund: u128,
+    pub token_amount_in_fund: u64,
     pub minted_lrt_mint: Pubkey,
     pub minted_lrt_amount: u64,
     pub lrt_amount_in_user_lrt_account: u64,
